+++ spec/dummy/config/routes.rb
@@ -6,6 +6,8 @@ Rails.application.routes.draw do
   get  "/notice", to: "pages#show_notice"
   get  "/alert",  to: "pages#show_alert"
   get  "/custom_flash", to: "pages#show_custom"
+  get  "/stacked_errors", to: "pages#show_stacked"
+  get  "/sequential_errors", to: "pages#show_sequential"
   get  "/confirm", to: "pages#confirm_page"
   get  "/chain",   to: "pages#chain_page"
   delete "/items/:id", to: "pages#destroy", as: :item